Fix following issues (1) In Migration select all not working with Assumption Master. Also restrict duplicate records to migrate. (2) Goals View: Implement loan calculation formulas with different options. (3) Report: Add chart for networth and goal description page. (4) Improve performance to display report.

+++ PlannerInfo/NonFinancialAssetInfo.cs
@@ -232,7 +232,7 @@ namespace FinancialPlannerClient.PlannerInfo
                     nonFinancialAsset.OtherHolderName = dr.Field<string>("OtherHolderName");
                     nonFinancialAsset.OtherHolderShare = int.Parse(dr.Field<string>("OtherHolderShare"));
                     nonFinancialAsset.MappedGoalId = int.Parse(dr.Field<string>("MappedGoalId"));
-                    nonFinancialAsset.AssetMappingShare = int.Parse(dr.Field<string>("AssetMappingShare"));
+                    nonFinancialAsset.AssetMappingShare = decimal.Parse(dr.Field<string>("AssetMappingShare"));
                     nonFinancialAsset.AssetRealisationYear = dr.Field<string>("AssetRealisationYear");
                     nonFinancialAsset.Description = dr.Field<string>("Description");
                     nonFinancialAsset.GrowthPercentage = decimal.Parse(dr["GrowthPercentage"].ToString());
